+++ ElvUI_LuckyoneUI/Media/SharedMedia.lua
@@ -7,10 +7,9 @@ LSM:Register('font','Alpha Echo', [[Interface\AddOns\ElvUI_LuckyoneUI\Media\Font
 LSM:Register('font','BeeLeave', [[Interface\AddOns\ElvUI_LuckyoneUI\Media\Fonts\BeeLeave.ttf]])
 LSM:Register('font','Dreamwood', [[Interface\AddOns\ElvUI_LuckyoneUI\Media\Fonts\Dreamwood.ttf]])
 LSM:Register('font','Handa', [[Interface\AddOns\ElvUI_LuckyoneUI\Media\Fonts\Handa.ttf]])
-LSM:Register('font','Kufam', [[Interface\AddOns\ElvUI_LuckyoneUI\Media\Fonts\Kufam.ttf]])
 LSM:Register('font','Meteora', [[Interface\AddOns\ElvUI_LuckyoneUI\Media\Fonts\Meteora.ttf]])
+LSM:Register('font','OrelegaOne', [[Interface\AddOns\ElvUI_LuckyoneUI\Media\Fonts\OrelegaOne-Regular.ttf]])
 LSM:Register('font','Roboto Light', [[Interface\AddOns\ElvUI_LuckyoneUI\Media\Fonts\Roboto Light.ttf]])
-LSM:Register('font','Teko', [[Interface\AddOns\ElvUI_LuckyoneUI\Media\Fonts\Teko.ttf]])
 
 -- Statusbar Textures
 LSM:Register('statusbar','Atrocity', [[Interface\AddOns\ElvUI_LuckyoneUI\Media\Statusbar\Atrocity.tga]])